Removed private mail
I have no such clue why is it so hard for codeberg(?) to identify the same user with different mails but it cannot handle the fact that i changed my mail address in my account. my registration is not new but looks like there is a configuration problem with this project and i cannot select my contribution mail address if it is not set as primary address. weblate can do this.

so please check the codeberg/weblate translation configurations and remove somehow my .hu mail. thanks.

Signed-off-by: ViBE <ViBE-HU@users.noreply.github.com>